Add files via upload
Partially fixed the -1 index error in the when interacting with the DataGridView.
When first time using the program, the index outputs an error when interacting with it, save the progress and reopening the program fixes the problem

Added a new feature using .json file that automatically saves the index even after the program's termination. Can continue operation in between sessions.